+++ GiganticEmu.Launcher/Services/GameLauncher.cs
@@ -3,81 +3,149 @@ using System.IO;
 using System.Threading.Tasks;
 using GiganticEmu.Shared;
 using Splat;
+using PeNet;
+using System.Linq;
+using System;
+using System.Collections.Generic;
 
 namespace GiganticEmu.Launcher;
 
 public class GameLauncher
 {
+    public record InteractionHandler
+    {
+        public Func<DialogContent, Task> OnError { get; init; } = async (_) => { };
+        public Func<DialogContent, Task<bool>> OnConfirm { get; init; } = async (_) => false;
+    }
+
     public static bool GameFound =>
         File.Exists(Path.Join(Locator.Current.RequireService<LauncherConfiguration>().Game, "Binaries", "Win64",
             "RxGame-Win64-Test.exe"));
 
-    public async Task RunGame(string? username = null, string? nickname = null, string? authToken = null)
+    public async Task RunGame(InteractionHandler interactions, string? username = null, string? nickname = null, string? authToken = null)
     {
         var config = Locator.Current.RequireService<LauncherConfiguration>();
         var settings = Locator.Current.RequireService<Settings>();
         var github = Locator.Current.RequireService<GitHub>();
         var path = Path.GetFullPath(Path.Join(config.Game, "Binaries", "Win64"));
 
-        if (!File.Exists(Path.Join(path, "RxGame-Win64-Test.exe")))
+        try
         {
-            // MessageBox.Show($"{Path.Join(path, "RxGame-Win64-Test.exe")} not found!", "File not found!", MessageBoxButton.OK, MessageBoxImage.Error);
-            return;
-        }
 
-        var giganticMetadata = FileVersionInfo.GetVersionInfo(Path.Join(path, "RxGame-Win64-Test.exe"));
-
-        var arcsdkMetadata = File.Exists(Path.Join(path, "ArcSDK.dll")) ? FileVersionInfo.GetVersionInfo(Path.Join(path, "ArcSDK.dll")) : null;
-
-        if (arcsdkMetadata is not { ProductName: "ArcSDK" } || arcsdkMetadata.ProductVersion == null)
-        {
-            // var result = MessageBox.Show("Unknown ArcSDK.dll version found.\nThis probably means this is your first time running MistforgeLauncher.\nMistforgeLauncher needs to replace the ArcSDK.dll with it's own version to continue.\n\nDo you want to continue?", "Unknown ArcSDK.dll version!", MessageBoxButton.YesNo, MessageBoxImage.Question);
-            // if (result != MessageBoxResult.Yes) return;
-
-            await github.DownloadFile(SemVer.ApplicationVersion, "ArcSDK.dll", path);
-        }
-        else
-        {
-            var arcSdkVersion = SemVer.Parse(arcsdkMetadata.ProductVersion);
-
-            if (SemVer.ApplicationVersion != arcSdkVersion)
+            if (!File.Exists(Path.Join(path, "RxGame-Win64-Test.exe")))
             {
-                // var result = MessageBox.Show("Mismatched ArcSDK.dll version detected. The version of MistforgeLauncher and ArcSDK.dll should match or problems can occur.\n\nShould MistforgeLauncher replace your ArcSDK.dll with the correct version?", "Mismatched ArcSDK.dll version!", MessageBoxButton.YesNo, MessageBoxImage.Question);
-                // if (result == MessageBoxResult.Yes)
-                // {
-                //     await github.DownloadFile(SemVer.ApplicationVersion, "ArcSDK.dll", path);
-                // }
+                await interactions.OnError(new DialogContent
+                {
+                    Title = "File not found!",
+                    Text = $"{Path.Join(path, "RxGame-Win64-Test.exe")} not found!",
+                });
+                return;
             }
+
+            var giganticMetadata = new PeFile(Path.Join(path, "RxGame-Win64-Test.exe"))?.Resources?.VsVersionInfo?.StringFileInfo?.StringTable?.FirstOrDefault();
+
+            var arcsdkMetadata = File.Exists(Path.Join(path, "ArcSDK.dll")) ? new PeFile(Path.Join(path, "ArcSDK.dll"))?.Resources?.VsVersionInfo?.StringFileInfo?.StringTable?.FirstOrDefault() : null;
+
+            if (arcsdkMetadata is not { ProductName: "ArcSDK" } || arcsdkMetadata.ProductVersion == null)
+            {
+                var result = await interactions.OnConfirm(new DialogContent
+                {
+                    Title = "Unknown ArcSDK.dll version!",
+                    Text = "Unknown ArcSDK.dll version found.\nThis probably means this is your first time running MistforgeLauncher.\nMistforgeLauncher needs to replace the ArcSDK.dll with it's own version to continue.\n\nDo you want to continue?"
+                });
+
+                if (!result)
+                {
+                    return;
+                }
+
+                await github.DownloadFile(SemVer.ApplicationVersion, "ArcSDK.dll", path);
+            }
+            else
+            {
+                var arcSdkVersion = SemVer.Parse(arcsdkMetadata.ProductVersion);
+
+                if (SemVer.ApplicationVersion != arcSdkVersion)
+                {
+                    var result = await interactions.OnConfirm(new DialogContent
+                    {
+                        Title = "Mismatched ArcSDK.dll version!",
+                        Text = "Mismatched ArcSDK.dll version detected. The version of MistforgeLauncher and ArcSDK.dll should match or problems can occur.\n\nShould MistforgeLauncher replace your ArcSDK.dll with the correct version?"
+                    });
+                    if (result)
+                    {
+                        await github.DownloadFile(SemVer.ApplicationVersion, "ArcSDK.dll", path);
+                    }
+                }
+            }
+
+            var commands = new List<string>();
+
+            var process = new Process();
+
+            if (PlatformUtils.IsLinux)
+            {
+                if (settings.LinuxEnableGameMode.Value)
+                {
+                    commands.Add("gamemoderun");
+                }
+
+                if (settings.LinuxEnableMangoHud.Value)
+                {
+                    commands.Add("mangohud");
+                }
+
+                commands.AddRange(settings.LinuxCompatiblityTool.Value switch
+                {
+                    Settings.CompatiblityTool.Proton => new[] { "proton", "run" },
+                    Settings.CompatiblityTool.Wine => new[] { "wine" },
+
+                });
+            }
+
+            commands.Add(Path.Join(path, "RxGame-Win64-Test.exe"));
+            commands.Add($"?name={nickname ?? "Offline"}");
+
+            commands.Add($"-ini:RxEngine:MotigaAuthIntegration.AuthUrlPrefix={config.Host}/,ArcIntegration.AuthUrlPrefix={config.Host}/");
+            commands.Add($"-log=GiganticEmu.Launcher.{username ?? "offline"}.log");
+
+            commands.Add($"-emu:nickname={nickname ?? "Offline"}");
+            commands.Add($"-emu:username={username ?? "Offline"}");
+            commands.Add($"-emu:auth_token={authToken ?? ""}");
+            commands.Add(@$"-emu:language={settings.GameLanguage.Value switch
+            {
+                Settings.Language.English => "INT",
+                Settings.Language.German => "DEU",
+                Settings.Language.French => "FRA",
+            }}");
+            commands.Add($"-emu:launch_code={(giganticMetadata?.ProductVersion == "1, 0, 16601, 0" ? 0 : 0xE0000019)}");
+
+
+            process.StartInfo = new ProcessStartInfo
+            {
+                FileName = commands.First()
+            };
+
+            foreach (var arg in commands.Skip(1))
+            {
+                process.StartInfo.ArgumentList.Add(arg);
+            }
+
+            process.Start();
+            await process.WaitForExitAsync();
         }
-
-        var process = new Process();
-        process.StartInfo = new ProcessStartInfo
+        catch (Exception ex)
         {
-            FileName = Path.Join(path, "RxGame-Win64-Test.exe")
-        };
-
-        process.StartInfo.ArgumentList.Add($"?name={nickname ?? "Offline"}");
-
-        process.StartInfo.ArgumentList.Add($"-ini:RxEngine:MotigaAuthIntegration.AuthUrlPrefix={config.Host}/,ArcIntegration.AuthUrlPrefix={config.Host}/");
-        process.StartInfo.ArgumentList.Add($"-log=GiganticEmu.Launcher.{username ?? "offline"}.log");
-
-        process.StartInfo.ArgumentList.Add($"-emu:nickname={nickname ?? "Offline"}");
-        process.StartInfo.ArgumentList.Add($"-emu:username={username ?? "Offline"}");
-        process.StartInfo.ArgumentList.Add($"-emu:auth_token={authToken ?? ""}");
-        process.StartInfo.ArgumentList.Add(@$"-emu:language={settings.GameLanguage.Value switch
-        {
-            Settings.Language.English => "INT",
-            Settings.Language.German => "DEU",
-            Settings.Language.French => "FRA",
-        }}");
-        process.StartInfo.ArgumentList.Add($"-emu:launch_code={(giganticMetadata.ProductBuildPart == 16601 ? 0 : 0xE0000019)}");
-
-        process.Start();
-        await process.WaitForExitAsync();
+            interactions?.OnError(new DialogContent
+            {
+                Title = "Error during startup!",
+                Text = ex.Message,
+            });
+        }
     }
 
-    public void StartGame(string? username = null, string? nickname = null, string? authToken = null)
+    public void StartGame(InteractionHandler interactions, string? username = null, string? nickname = null, string? authToken = null)
     {
-        _ = Task.Run(async () => await RunGame(username, nickname, authToken));
+        _ = Task.Run(async () => await RunGame(interactions, username, nickname, authToken));
     }
 }