+++ GiganticEmu.Launcher/ViewModels/DialogContent.cs
@@ -1,0 +1,8 @@
+
+namespace GiganticEmu.Launcher;
+
+public record DialogContent
+{
+    public string Title { get; init; } = default!;
+    public string Text { get; init; } = default!;
+}