+++ GiganticEmu.Launcher/ViewModels/IDialogViewModel.cs
@@ -5,6 +5,5 @@ namespace GiganticEmu.Launcher;
 
 public interface IDialogViewModel<T>
 {
-
     public ReactiveCommand<Unit, T?> Show { get; }
 }