+++ GiganticEmu.Launcher/ViewModels/MainContainerViewModel.cs
@@ -2,6 +2,7 @@ using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
 using Splat;
 using System.Reactive;
+using System.Reactive.Linq;
 using System.Threading.Tasks;
 
 namespace GiganticEmu.Launcher;
@@ -16,6 +17,10 @@ public class MainContainerViewModel : ReactiveObject
 
     public ReactiveCommand<Unit, Unit> StartGame { get; }
 
+    public Interaction<DialogContent, Unit> OnError { get; } = new();
+
+    public Interaction<DialogContent, bool> OnConfirm { get; } = new();
+
     public MainContainerViewModel()
     {
         StartGame = ReactiveCommand.CreateFromTask(DoStartGame);
@@ -24,6 +29,10 @@ public class MainContainerViewModel : ReactiveObject
     private async Task DoStartGame()
     {
         var launcher = Locator.Current.RequireService<GameLauncher>();
-        launcher.StartGame(User?.UserName, User?.UserName, User?.AuthToken);
+        launcher.StartGame(new GameLauncher.InteractionHandler
+        {
+            OnConfirm = async content => await OnConfirm.Handle(content),
+            OnError = async content => await OnError.Handle(content),
+        }, User?.UserName, User?.UserName, User?.AuthToken);
     }
 }