+++ GiganticEmu.Launcher/Views/Login/Login.axaml.cs
@@ -4,6 +4,7 @@ using Avalonia.ReactiveUI;
 using System.Reactive.Linq;
 using System.Reactive.Disposables;
 using Avalonia.Controls;
+using System.Reactive;
 
 namespace GiganticEmu.Launcher;
 
@@ -60,6 +61,41 @@ public partial class Login : ReactiveUserControl<LoginViewModel>
                 })
                 .Subscribe()
                 .DisposeWith(disposables);
+
+            ViewModel.OnConfirm.RegisterHandler(async interaction =>
+            {
+                var result = await this.ShowAlertDialog(async dialog =>
+                {
+                    dialog.Title = interaction.Input.Title;
+                    dialog.Text = interaction.Input.Text;
+                    dialog.Icon = "HelpCircleOutline";
+
+                    dialog.Buttons.Edit(buttons =>
+                    {
+                        buttons.Add(new AlertDialogViewModel.Button("Yes", "yes"));
+                        buttons.Add(new AlertDialogViewModel.Button("No", "no"));
+                    });
+                });
+
+                interaction.SetOutput(result == "yes");
+            });
+
+            ViewModel.OnError.RegisterHandler(async interaction =>
+            {
+                var result = await this.ShowAlertDialog(async dialog =>
+                {
+                    dialog.Title = interaction.Input.Title;
+                    dialog.Text = interaction.Input.Text;
+                    dialog.Icon = "AlertCircleOutline";
+
+                    dialog.Buttons.Edit(buttons =>
+                    {
+                        buttons.Add(new AlertDialogViewModel.Button("Ok", "ok"));
+                    });
+                });
+
+                interaction.SetOutput(Unit.Default);
+            });
         });
     }
 }