+++ GiganticEmu.Launcher/Views/Main/MainContainer.axaml.cs
@@ -20,6 +20,41 @@ public partial class MainContainer : ReactiveUserControl<MainContainerViewModel>
                 .Select(x => Unit.Default)
                 .InvokeCommand(this, x => x.ViewModel!.StartGame)
                 .DisposeWith(disposables);
+
+            ViewModel.OnConfirm.RegisterHandler(async interaction =>
+            {
+                var result = await this.ShowAlertDialog(async dialog =>
+                {
+                    dialog.Title = interaction.Input.Title;
+                    dialog.Text = interaction.Input.Text;
+                    dialog.Icon = "HelpCircleOutline";
+
+                    dialog.Buttons.Edit(buttons =>
+                    {
+                        buttons.Add(new AlertDialogViewModel.Button("Yes", "yes"));
+                        buttons.Add(new AlertDialogViewModel.Button("No", "no"));
+                    });
+                });
+
+                interaction.SetOutput(result == "yes");
+            });
+
+            ViewModel.OnError.RegisterHandler(async interaction =>
+            {
+                var result = await this.ShowAlertDialog(async dialog =>
+                {
+                    dialog.Title = interaction.Input.Title;
+                    dialog.Text = interaction.Input.Text;
+                    dialog.Icon = "AlertCircleOutline";
+
+                    dialog.Buttons.Edit(buttons =>
+                    {
+                        buttons.Add(new AlertDialogViewModel.Button("Ok", "ok"));
+                    });
+                });
+
+                interaction.SetOutput(Unit.Default);
+            });
         });
     }
 }